added create_store process for wizard installation
