+++ view/adminhtml/templates/setup.phtml
@@ -107,9 +107,7 @@
                 "loginEndpoint": "<?= $this->helper('Doofinder\Feed\Helper\StoreConfig')->getLoginEndpoint() ?>",
                 "paramsPopup": "<?= $block->getParamsPopup() ?>",
                 "checkAPIKeyUrl": "<?= $block->getCheckAPIKeyUrl() ?>",
-                "createSearchEnginesUrl": "<?= $block->getCreateSearchEnginesUrl() ?>",
-                "createIndicesUrl": "<?= $block->getCreateIndicesUrl() ?>",
-                "processSearchEnginesUrl": "<?= $block->getProcessSearchEnginesUrl() ?>",
+                "createStoreUrl": "<?= $block->getCreateStoreUrl() ?>",
                 "createDisplayLayersUrl": "<?= $block->getCreateDisplayLayersUrl() ?>"
             }
         }
